Fix region resolution
