+++ quickwit-indexing/src/source/kinesis/kinesis_source.rs
@@ -25,6 +25,7 @@ use anyhow::Context;
 use async_trait::async_trait;
 use itertools::Itertools;
 use quickwit_actors::{ActorExitStatus, Mailbox};
+use quickwit_aws::region::sniff_s3_region_and_cache;
 use quickwit_aws::retry::RetryParams;
 use quickwit_config::{KinesisSourceParams, RegionOrEndpoint};
 use quickwit_metastore::checkpoint::{
@@ -340,16 +341,23 @@ impl Source for KinesisSource {
 }
 
 pub(super) fn get_region(region_or_endpoint: Option<RegionOrEndpoint>) -> anyhow::Result<Region> {
-    match region_or_endpoint {
-        Some(RegionOrEndpoint::Endpoint(endpoint)) => Ok(Region::Custom {
+    if let Some(RegionOrEndpoint::Endpoint(endpoint)) = region_or_endpoint {
+        return Ok(Region::Custom {
             name: "Custom".to_string(),
             endpoint,
-        }),
-        Some(RegionOrEndpoint::Region(region)) => region
-            .parse()
-            .with_context(|| format!("Failed to parse region: `{}`", region)),
-        None => Ok(Region::default()),
+        });
     }
+
+    // TODO: Change this to let_chain when 1.64 lands. (this is ugly)
+    if let Some(RegionOrEndpoint::Region(region)) = region_or_endpoint {
+        match region.parse() {
+            Ok(region_val) => return Ok(region_val),
+            Err(_) => {}
+        }
+    }
+
+    sniff_s3_region_and_cache() //< We fallback to S3 region if parsing from `region_or_endpoint`
+                                //< fails
 }
 
 #[cfg(all(test, feature = "kinesis-localstack-tests"))]
@@ -375,6 +383,35 @@ mod tests {
         }
         merged_batch.docs.sort();
         Ok(merged_batch)
+    }
+
+    #[test]
+    fn test_kinesis_region_resolution() {
+        {
+            let region_or_endpoint = Some(RegionOrEndpoint::Endpoint(
+                "mycustomendpoint.quickwit".to_string(),
+            ));
+            let region = get_region(region_or_endpoint).unwrap();
+            assert_eq!(
+                Region::Custom {
+                    name: "Custom".to_string(),
+                    endpoint: "mycustomendpoint.quickwit".to_string()
+                },
+                region
+            );
+        }
+
+        {
+            let region_or_endpoint = Some(RegionOrEndpoint::Region("us-east-1".to_string()));
+            let region = get_region(region_or_endpoint).unwrap();
+            assert_eq!(Region::UsEast1, region);
+        }
+
+        {
+            let region_or_endpoint = Some(RegionOrEndpoint::Region("quickwit-hq-1".to_string()));
+            let region = get_region(region_or_endpoint).unwrap();
+            assert_eq!(sniff_s3_region_and_cache().unwrap(), region);
+        }
     }
 
     #[tokio::test]
