+++ quickwit-storage/Cargo.toml
@@ -14,7 +14,6 @@ anyhow = "1"
 async-trait = "0.1"
 base64 = "0.13"
 bytes = "1"
-ec2_instance_metadata = "0.3"
 fnv = "1.0.7"
 futures = "0.3"
 lru = "0.7"
